+++ src/containerapp/HISTORY.rst
@@ -8,6 +8,7 @@ upcoming
 * 'az containerapp connected-env certificate upload/remove': Support `--no-wait`.
 * 'az containerapp connected-env dapr-component set/remove': Support `--no-wait`.
 * 'az containerapp connected-env storage set/remove': Support `--no-wait`.
+* 'az containerapp env http-route-config': Add commands for the http-route-config feature area.
 
 1.1.0b1
 ++++++
@@ -16,7 +17,7 @@ upcoming
 * 'az containerapp create/update': `--yaml` support property pollingInterval and cooldownPeriod
 * 'az containerapp session code-interpreter upload-file/list-files/show-file-content/show-file-metadata/delete-file': Support `--path` to specify the path of code interpreter session file resource
 * 'az containerapp session code-interpreter': Update response payload format for api-version 2024-10-02-preview
-* 'az containerapp env maintenance-config add/update/list/remove': Support environment maintenance config management 
+* 'az containerapp env maintenance-config add/update/list/remove': Support environment maintenance config management
 * 'az containerapp sessionpool create': Support managed identity when create session pool with --mi-system-assigned --mi-user-assigned
 
 1.0.0b4